@yamaguchi1024 2018-09-06
@vgvassilev 
I think we don't have to do Evaluate each time, rather declaring RuntimePrintValue is kind for users. Also as I wrote in a commit message, this will be a no-op with Cling modules