@Axel-Naumann 2018-09-06
IMHO we should fix the invocation of printValue in RDF to use the official cling interface, instead of making cling slower to start up.