@yamaguchi1024 2018-09-19
If printValue was not intended to use in RDataFrame etc, what the use case of this?