@yamaguchi1024 2018-09-19
Also it doesn't change performance at all with cling modules.
Without this patch:
```
Processing /home/yuka/CERN/ROOT/memory.C...
 cpu  time = 0.233852 seconds
 sys  time = 0.092215 seconds
 res  memory = 210.836 Mbytes
 vir  memory = 346.195 Mbytes
```

With this patch:
```
Processing /home/yuka/CERN/ROOT/memory.C...                                                                                                                                                                        
 cpu  time = 0.248781 seconds                                                                                                                                                                                      
 sys  time = 0.067224 seconds                                                                                                                                                                                      
 res  memory = 210.258 Mbytes                                                                                                                                                                                      
 vir  memory = 345.773 Mbytes   
```

So I think there is not downside for this