@yamaguchi1024 2018-09-24
#2664 is a superior implementation of this